feat(m2.10): extract Communications domain bounded context with zero functional impact
