+++ core/src/domains/communications/application/services/ChatService.ts
@@ -1,0 +1,5 @@
+export * from './chat/ChatUtils';
+export * from './chat/ChatConversationService';
+export * from './chat/ChatMessageService';
+export * from './chat/ChatReportService';
+export * from './chat/ChatAdminService';
