+++ core/src/domains/communications/index.ts
@@ -1,0 +1,4 @@
+export * from './application/services/ContactService';
+export * from './application/services/ContactRevealService';
+export * from './application/services/ChatService';
+export * from './application/services/ChatAvailabilityService';
