+++ core/src/services/ChatService.ts
@@ -1,5 +1,1 @@
-export * from './chat/ChatUtils';
-export * from './chat/ChatConversationService';
-export * from './chat/ChatMessageService';
-export * from './chat/ChatReportService';
-export * from './chat/ChatAdminService';
+export * from '../domains/communications/application/services/ChatService';
